+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Fix #113: Adds DOCKER_API_VERSION in env docker output @navidshaikh
+- Adds SPEC file version 0.0.4 of the plugin @navidshaikh
 
 ## v0.0.4 Mar 14, 2016
 - Fix #101: vagrant-service-manager version 0.0.4 release @navidshaikh