+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix #65: Adds --script-readable option for `env openshift` command @navidshaikh
 - Fix #80: Check for correct TLS certs pair @budhrg
 - Fix #113: Adds DOCKER_API_VERSION in env docker output @navidshaikh
+- Adds SPEC file version 0.0.4 of the plugin @navidshaikh
 
 ## v0.0.4 Mar 14, 2016
 - Fix #101: vagrant-service-manager version 0.0.4 release @navidshaikh